refactor: Separate lexer utils and string utils
